+++ server/global/csv_reader/ParserSpecifications.js
@@ -7,11 +7,14 @@ export class ColumnSpecifications {
    * @param {string} title The column title
    * @param {string} property The property corresponding
    * @param {number} type The columns type (unique, hierarchical, multivalued)
+   * @param {{key? : boolean, maxlength? : number}} options Optional options
    */
-  constructor(title, property, type) {
+  constructor(title, property, type, { key = false, maxlength = null } = {}) {
     this.title = title;
     this.property = property;
     this.type = type;
+    this.key = key;
+    this.maxlength = maxlength;
   }
   isUnique() {
     return this.type === ColumnSpecifications.UNIQUE;
@@ -49,17 +52,4 @@ ColumnSpecifications.HIERARCHICAL = 1;
 ColumnSpecifications.UNIQUE = 2;
 ColumnSpecifications.MULTI_VALUED = 3;
 
-const columns = [
-  new ColumnSpecifications("DCI", "dci", ColumnSpecifications.UNIQUE),
-  new ColumnSpecifications("FORMULE_CHIMIQUE", "skeletal_formule", ColumnSpecifications.UNIQUE),
-  new ColumnSpecifications("SYSTEME_n", "systems", ColumnSpecifications.HIERARCHICAL),
-  new ColumnSpecifications("CLASSE_PHARMA_n", "classes", ColumnSpecifications.HIERARCHICAL),
-  new ColumnSpecifications("MTE", "ntr", ColumnSpecifications.UNIQUE),
-  new ColumnSpecifications("INTERACTION", "interactions", ColumnSpecifications.MULTI_VALUED),
-  new ColumnSpecifications("INDICATION", "indications", ColumnSpecifications.MULTI_VALUED),
-  new ColumnSpecifications("EFFET_INDESIRABLE", "side_effects", ColumnSpecifications.MULTI_VALUED),
-  new ColumnSpecifications("NIVEAU_DEBUTANT", "level_easy", ColumnSpecifications.UNIQUE),
-  new ColumnSpecifications("NIVEAU_EXPERT", "level_hard", ColumnSpecifications.UNIQUE),
-];
-
-export default { columns };
+export default ColumnSpecifications;
