+++ server/test/csv_parser/parser.test.js
@@ -4,11 +4,11 @@ import mocha from "mocha";
 import fs from "fs";
 import deepEqualAnyOrder from "deep-equal-in-any-order";
 
-import { parseCSV } from "../../global/data_importer/csv_parser/Parser.js";
+import { parseMoleculesFromCsv } from "../../global/molecules_parser/Parser.js";
 import { expectations } from "./expectations.js";
 // eslint-disable-next-line no-unused-vars
-import { ClassificationNode } from "../../global/data_importer/csv_parser/MoleculesClassification.js";
-import { HeaderErrors } from "../../global/data_importer/csv_parser/HeaderChecker.js";
+import { ClassificationNode } from "../../global/molecules_parser/MoleculesClassification.js";
+import { HeaderErrors } from "../../global/csv_reader/HeaderChecker.js";
 
 chai.use(deepEqualAnyOrder);
 const { before } = mocha;
@@ -57,7 +57,7 @@ describe("Test if values are well imported", function () {
       let data;
 
       before("Import data", (done) => {
-        parseCSV(path.resolve(filesFolderPath, file.name))
+        parseMoleculesFromCsv(path.resolve(filesFolderPath, file.name))
           .then((json) => {
             data = JSON.parse(json);
             done();
@@ -225,7 +225,7 @@ describe("Tests for errors occurred while parsing an incorrectly formatted file"
   for (const file of bad_files) {
     it(`File : ${file.name}`, async () => {
       try {
-        const json = await parseCSV(path.resolve(badFilesFolderPath, file.name));
+        const json = await parseMoleculesFromCsv(path.resolve(badFilesFolderPath, file.name));
         expect(json === null, "The parsing is not supposed to pass").be.true;
       } catch (errors) {
         if (!HeaderErrors.isInstance(errors)) {
